test(jwt): consolidate issuer tests into jwt_test.go
Move the issuer-enforcement tests and their helpers from jwt_issuer_test.go
into jwt_test.go, as requested in review, keeping all JWT identity tests in a
single file.
